requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.5
python-multipart==0.0.20
pandas==2.2.3
openpyxl==3.1.5
PyMuPDF==1.25.1
python-docx==1.1.2
rapidfuzz==3.11.0
Unidecode==1.3.8
pytest==8.3.4
